Fixed issue with getting sources
