+++ anchor-api/lib/controllers/getArticles.ts
@@ -37,13 +37,3 @@ export function getArticles(req: express.Request, res: express.Response) {
         res.status(200).send(results);
       });
   }
-
-  export function getSources(req: express.Request, res: express.Response) {
-    Article.find().distinct('publisher', (error: any, results: any) => {
-      if (error) {
-        res.status(403).send(error);
-        throw error;
-      }
-      res.status(200).send(results);
-    })
-  }
